refactor: :bulb: heartbeat detection components and logic
