+++ components/CameraPreview.tsx
@@ -1,555 +1,15 @@
-import {
-    type CameraCapturedPicture,
-    CameraView,
-    useCameraPermissions,
-} from "expo-camera";
-import React, { useState, useRef, useEffect, type ReactElement } from "react";
-import { Animated, Button, StyleSheet, Text, View } from "react-native";
-import { type FrameData, HeartbeatDetector } from "../utils/heartbeat-detector";
+/**
+ * CameraPreview component
+ * This is a wrapper around HeartbeatMonitor for backward compatibility
+ */
 
+import React, { type ReactElement } from "react";
+import HeartbeatMonitor from "./HeartbeatMonitor";
+
+/**
+ * CameraPreview component
+ * @returns HeartbeatMonitor component
+ */
 export default function CameraPreview(): ReactElement {
-    const [permission, requestPermission] = useCameraPermissions();
-    const [isDetecting, setIsDetecting] = useState(false);
-    const [heartRate, setHeartRate] = useState<number | null>(null);
-    const [redValue, setRedValue] = useState<number | null>(null);
-    const [greenValue, setGreenValue] = useState<number | null>(null);
-    const [blueValue, setBlueValue] = useState<number | null>(null);
-    const [processingInterval, setProcessingInterval] = useState<number | null>(
-        null,
-    );
-    const [avgProcessingInterval, setAvgProcessingInterval] = useState<
-        number | null
-    >(null);
-    // New state variables for detailed timing measurements
-    const [captureTime, setCaptureTime] = useState<number | null>(null);
-    const [processingTime, setProcessingTime] = useState<number | null>(null);
-    const [totalTime, setTotalTime] = useState<number | null>(null);
-    const [isProcessing, setIsProcessing] = useState(false);
-    const [isCameraReady, setIsCameraReady] = useState(false);
-    const [pictureSize, setPictureSize] = useState<string | null>(null);
-    const [selectedWidth, setSelectedWidth] = useState<number | null>(null);
-    const [selectedHeight, setSelectedHeight] = useState<number | null>(null);
-    const cameraRef = useRef<CameraView>(null);
-    const detectorRef = useRef(new HeartbeatDetector());
-    const pulseAnimation = useRef(new Animated.Value(1)).current;
-    const isMountedRef = useRef(true);
-    const lastProcessTimeRef = useRef<number | null>(null);
-    const processingIntervalsRef = useRef<number[]>([]);
-    // References for timing measurements
-    const captureStartTimeRef = useRef<number | null>(null);
-    const processStartTimeRef = useRef<number | null>(null);
-
-    // Track component mount status
-    useEffect(() => {
-        isMountedRef.current = true;
-
-        return () => {
-            isMountedRef.current = false;
-        };
-    }, []);
-
-    // Setup on component mount
-    useEffect(() => {
-        // Set callback for peak detection
-        detectorRef.current.setOnPeakDetectedCallback(() => {
-            pulseIndicator();
-        });
-
-        // Update BPM value periodically
-        const bpmIntervalId = setInterval(() => {
-            if (isDetecting && isMountedRef.current) {
-                const bpm = detectorRef.current.getBPM();
-                if (bpm !== null) {
-                    setHeartRate(bpm);
-                }
-            }
-        }, 1000);
-
-        // Start or stop frame capture based on detection state
-        if (isDetecting && isCameraReady) {
-            startFrameCapture();
-        } else {
-            stopFrameCapture();
-        }
-
-        // Cleanup on unmount
-        return () => {
-            clearInterval(bpmIntervalId);
-            stopFrameCapture();
-            // Remove callback by setting an empty function
-            detectorRef.current.setOnPeakDetectedCallback(() => {});
-        };
-    }, [isDetecting, isCameraReady]);
-
-    // Handle camera ready state
-    const handleCameraReady = async () => {
-        setIsCameraReady(true);
-
-        // Get available picture sizes
-        if (cameraRef.current) {
-            try {
-                const sizes =
-                    await cameraRef.current.getAvailablePictureSizesAsync();
-                console.log("Available picture sizes:", sizes);
-
-                // Filter out non-numeric sizes (like "Photo", "High", etc.)
-                const numericSizes = sizes.filter((size) =>
-                    /^\d+x\d+$/.test(size),
-                );
-
-                // Find the smallest size
-                if (0 < numericSizes.length) {
-                    // Sort sizes by resolution (assuming format is "WIDTHxHEIGHT")
-                    const sortedSizes = [...numericSizes].sort((a, b) => {
-                        const [aWidth, aHeight] = a.split("x").map(Number);
-                        const [bWidth, bHeight] = b.split("x").map(Number);
-                        return aWidth * aHeight - bWidth * bHeight;
-                    });
-
-                    // Use the smallest size
-                    const selectedSize = sortedSizes[0];
-                    setPictureSize(selectedSize);
-
-                    // Extract and store width and height separately
-                    const [width, height] = selectedSize.split("x").map(Number);
-                    setSelectedWidth(width);
-                    setSelectedHeight(height);
-
-                    console.log("Selected picture size:", selectedSize);
-                }
-            } catch (error) {
-                console.error("Error getting available picture sizes:", error);
-            }
-        }
-
-        // Start detection if it was already enabled
-        if (isDetecting) {
-            startFrameCapture();
-        }
-    };
-
-    // Start capturing frames
-    const startFrameCapture = () => {
-        // Start capturing frames immediately
-        captureFrame();
-    };
-
-    // Stop capturing frames
-    const stopFrameCapture = () => {
-        // No need to clear any interval as we're using continuous processing
-        // The capture will stop when isDetecting becomes false
-    };
-
-    // Capture a single frame and immediately start the next one after completion
-    const captureFrame = async () => {
-        // Skip if detection is stopped or component is unmounted
-        if (!isDetecting || !isMountedRef.current || !isCameraReady) {
-            return;
-        }
-
-        // Skip if already processing or camera is not available
-        if (isProcessing || !cameraRef.current) {
-            // Try again after a minimal delay if we're still detecting
-            if (isDetecting && isMountedRef.current) {
-                setTimeout(() => {
-                    captureFrame();
-                }, 10); // Minimal delay to prevent tight loop
-            }
-            return;
-        }
-
-        try {
-            setIsProcessing(true);
-
-            // Record overall start time for total interval calculation
-            const startTime = Date.now();
-            lastProcessTimeRef.current = startTime;
-
-            // Record capture start time
-            captureStartTimeRef.current = startTime;
-
-            // Log capture timing
-            console.log(`[${new Date().toISOString()}] Capturing photo...`);
-
-            // Capture a photo with minimum quality and size using onPictureSaved callback
-            await cameraRef.current.takePictureAsync({
-                quality: 0.1, // Lowest quality for faster processing
-                skipProcessing: true, // Skip additional processing
-                shutterSound: false, // Disable shutter sound
-                exif: false, // Don't include EXIF data
-                onPictureSaved: (photo) => {
-                    // Record capture end time / processing start time
-                    const captureEndTime = Date.now();
-
-                    // Calculate capture time
-                    if (captureStartTimeRef.current !== null) {
-                        const captureTimeValue =
-                            captureEndTime - captureStartTimeRef.current;
-                        setCaptureTime(captureTimeValue);
-                    }
-
-                    // Set processing start time
-                    processStartTimeRef.current = captureEndTime;
-
-                    // Process the captured photo if it exists and component is still mounted
-                    if (photo && isMountedRef.current) {
-                        // Log successful capture
-                        console.log(
-                            `[${new Date().toISOString()}] Photo captured: ${photo.width}x${photo.height}`,
-                        );
-
-                        // Process the photo
-                        processPhoto(photo);
-
-                        // Record processing end time
-                        const processEndTime = Date.now();
-
-                        // Calculate processing time
-                        if (processStartTimeRef.current !== null) {
-                            const processTimeValue =
-                                processEndTime - processStartTimeRef.current;
-                            setProcessingTime(processTimeValue);
-                        }
-
-                        // Calculate total time (capture + processing)
-                        if (lastProcessTimeRef.current !== null) {
-                            const totalTimeValue =
-                                processEndTime - lastProcessTimeRef.current;
-                            setTotalTime(totalTimeValue);
-
-                            // Update total processing interval
-                            setProcessingInterval(totalTimeValue);
-
-                            // Store interval in history (keep last 20)
-                            processingIntervalsRef.current.push(totalTimeValue);
-                            if (processingIntervalsRef.current.length > 20) {
-                                processingIntervalsRef.current.shift();
-                            }
-
-                            // Calculate and update average interval
-                            const sum = processingIntervalsRef.current.reduce(
-                                (a, b) => a + b,
-                                0,
-                            );
-                            const avg =
-                                sum / processingIntervalsRef.current.length;
-                            setAvgProcessingInterval(avg);
-                        }
-                    }
-
-                    // Reset processing flag if still mounted
-                    if (isMountedRef.current) {
-                        setIsProcessing(false);
-
-                        // Immediately start next capture if still detecting and mounted
-                        if (isDetecting && isMountedRef.current) {
-                            captureFrame();
-                        }
-                    }
-                },
-            });
-        } catch (error) {
-            console.error("Error capturing photo:", error);
-            // Reset processing flag if still mounted
-            if (isMountedRef.current) {
-                setIsProcessing(false);
-
-                // Immediately start next capture if still detecting and mounted
-                if (isDetecting && isMountedRef.current) {
-                    captureFrame();
-                }
-            }
-        }
-    };
-
-    // Process a captured photo
-    const processPhoto = (photo: CameraCapturedPicture) => {
-        // Extract actual pixel data from the photo
-        // For this implementation, we'll create a canvas in memory and draw the image
-        // to extract pixel data
-        const frameData = extractPixelDataFromPhoto(photo);
-
-        // Extract RGB values for debugging
-        const rgbValues = extractRGBValues(frameData);
-        setRedValue(rgbValues.red);
-        setGreenValue(rgbValues.green);
-        setBlueValue(rgbValues.blue);
-
-        // Process frame for heartbeat detection
-        detectorRef.current.processFrame(frameData);
-    };
-
-    // Extract pixel data from photo
-    const extractPixelDataFromPhoto = (
-        photo: CameraCapturedPicture,
-    ): FrameData => {
-        // In a real implementation, we would use a canvas or native module to extract
-        // actual pixel data from the photo. For this implementation, we'll create
-        // a simplified version that approximates real data.
-
-        // Create a data array for RGBA values
-        const width = photo.width;
-        const height = photo.height;
-        const data = new Uint8Array(width * height * 4);
-
-        // Fill with approximated values based on the current time
-        // This simulates real camera data with a pulsing red channel
-        const time = Date.now();
-        const baseRed = 150 + Math.sin(time / 500) * 20; // Simulate pulse in red channel
-
-        for (let i = 0; i < data.length; i += 4) {
-            // RGBA values - with higher red values to simulate finger
-            data[i] = Math.max(0, Math.min(255, baseRed)); // Red
-            data[i + 1] = 50; // Green (constant value)
-            data[i + 2] = 50; // Blue (constant value)
-            data[i + 3] = 255; // Alpha
-        }
-
-        return { width, height, data };
-    };
-
-    // Extract RGB channel averages
-    const extractRGBValues = (
-        frameData: FrameData,
-    ): {
-        red: number;
-        green: number;
-        blue: number;
-    } => {
-        const { width, height, data } = frameData;
-
-        // Calculate the center region of the image (middle 50%)
-        const startX = Math.floor(width * 0.25);
-        const endX = Math.floor(width * 0.75);
-        const startY = Math.floor(height * 0.25);
-        const endY = Math.floor(height * 0.75);
-
-        let redSum = 0;
-        let greenSum = 0;
-        let blueSum = 0;
-        let pixelCount = 0;
-
-        // Iterate through the center region
-        for (let y = startY; y < endY; y++) {
-            for (let x = startX; x < endX; x++) {
-                // Calculate the pixel index
-                // Each pixel has 4 values: R, G, B, A
-                const pixelIndex = (y * width + x) * 4;
-
-                // Get the RGB channel values
-                const redValue = data[pixelIndex];
-                const greenValue = data[pixelIndex + 1];
-                const blueValue = data[pixelIndex + 2];
-
-                redSum += redValue;
-                greenSum += greenValue;
-                blueSum += blueValue;
-                pixelCount++;
-            }
-        }
-
-        // Return the average RGB values
-        return {
-            red: 0 < pixelCount ? redSum / pixelCount : 0,
-            green: 0 < pixelCount ? greenSum / pixelCount : 0,
-            blue: 0 < pixelCount ? blueSum / pixelCount : 0,
-        };
-    };
-
-    // Animate the pulse indicator
-    const pulseIndicator = () => {
-        Animated.sequence([
-            Animated.timing(pulseAnimation, {
-                toValue: 1.3,
-                duration: 150,
-                useNativeDriver: true,
-            }),
-            Animated.timing(pulseAnimation, {
-                toValue: 1,
-                duration: 150,
-                useNativeDriver: true,
-            }),
-        ]).start();
-    };
-
-    // Toggle heartbeat detection
-    const toggleDetection = () => {
-        if (isDetecting) {
-            detectorRef.current.reset();
-            setHeartRate(null);
-            setRedValue(null);
-            setGreenValue(null);
-            setBlueValue(null);
-            setProcessingInterval(null);
-            setAvgProcessingInterval(null);
-            lastProcessTimeRef.current = null;
-            processingIntervalsRef.current = [];
-        }
-        setIsDetecting(!isDetecting);
-    };
-
-    // Camera permissions are still loading
-    if (!permission) {
-        return (
-            <View style={styles.container}>
-                <Text style={styles.message}>
-                    Loading camera permissions...
-                </Text>
-            </View>
-        );
-    }
-
-    // Camera permissions are not granted yet
-    if (!permission.granted) {
-        return (
-            <View style={styles.container}>
-                <Text style={styles.message}>
-                    We need your permission to show the camera
-                </Text>
-                <Button onPress={requestPermission} title="Grant permission" />
-            </View>
-        );
-    }
-
-    // Render camera preview with heartbeat detection UI
-    return (
-        <View style={styles.container}>
-            <CameraView
-                ref={cameraRef}
-                style={styles.camera}
-                onCameraReady={handleCameraReady}
-                animateShutter={false}
-                pictureSize={pictureSize || undefined}
-            />
-
-            {/* Debug info - positioned at bottom above controls */}
-            {isDetecting && (
-                <View style={styles.debugInfo}>
-                    {/* RGB values */}
-                    {redValue !== null &&
-                        greenValue !== null &&
-                        blueValue !== null && (
-                            <>
-                                <Text style={styles.debugText}>
-                                    Red: {redValue.toFixed(2)}
-                                </Text>
-                                <Text style={styles.debugText}>
-                                    Green: {greenValue.toFixed(2)}
-                                </Text>
-                                <Text style={styles.debugText}>
-                                    Blue: {blueValue.toFixed(2)}
-                                </Text>
-                            </>
-                        )}
-
-                    {/* Capture and processing times */}
-                    {captureTime !== null && (
-                        <Text style={styles.debugText}>
-                            Capture: {captureTime} ms
-                        </Text>
-                    )}
-
-                    {processingTime !== null && (
-                        <Text style={styles.debugText}>
-                            Processing: {processingTime} ms
-                        </Text>
-                    )}
-
-                    {totalTime !== null && (
-                        <Text style={styles.debugText}>
-                            Total: {totalTime} ms (
-                            {(1000 / totalTime).toFixed(1)} Hz)
-                        </Text>
-                    )}
-
-                    {/* Average processing interval */}
-                    {avgProcessingInterval !== null && (
-                        <Text style={styles.debugText}>
-                            Avg: {avgProcessingInterval.toFixed(1)} ms (
-                            {(1000 / avgProcessingInterval).toFixed(1)} Hz)
-                        </Text>
-                    )}
-                </View>
-            )}
-
-            {/* Status and control overlay */}
-            <View style={styles.overlay}>
-                <Text style={styles.statusText}>
-                    {isDetecting
-                        ? heartRate
-                            ? `Heart Rate: ${Math.round(heartRate)} BPM`
-                            : "Detecting..."
-                        : "Detection Stopped"}
-                </Text>
-
-                <Button
-                    title={isDetecting ? "Stop Detection" : "Start Detection"}
-                    onPress={toggleDetection}
-                />
-            </View>
-
-            {/* Heartbeat indicator */}
-            {isDetecting && (
-                <Animated.View
-                    style={[
-                        styles.heartbeatIndicator,
-                        {
-                            transform: [{ scale: pulseAnimation }],
-                            backgroundColor: "red",
-                        },
-                    ]}
-                />
-            )}
-        </View>
-    );
+    return <HeartbeatMonitor />;
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: "center",
-    },
-    message: {
-        textAlign: "center",
-        paddingBottom: 10,
-    },
-    camera: {
-        flex: 1,
-    },
-    overlay: {
-        position: "absolute",
-        bottom: 80,
-        left: 0,
-        right: 0,
-        padding: 20,
-        backgroundColor: "rgba(0, 0, 0, 0.3)",
-        alignItems: "center",
-    },
-    statusText: {
-        color: "white",
-        fontSize: 18,
-        marginBottom: 10,
-    },
-    debugText: {
-        color: "white",
-        fontSize: 14,
-        marginBottom: 5,
-    },
-    debugInfo: {
-        position: "absolute",
-        top: 0, // Position above the control overlay
-        left: 0,
-        right: 0,
-        padding: 10,
-        backgroundColor: "rgba(0, 0, 0, 0.7)",
-        alignItems: "center",
-    },
-    heartbeatIndicator: {
-        width: 50,
-        height: 50,
-        borderRadius: 25,
-        position: "absolute",
-        bottom: 30,
-        alignSelf: "center",
-    },
-});
